get hero's

+++ src/main/java/com/protectors/app/protectorsservice/repository/SuperheroRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SuperheroRepository extends JpaRepository<Superhero, Integer> {
+public interface SuperheroRepository extends JpaRepository<Superhero, Long> {
 }
